Add a simple Admin UI
This commit adds a simple admin UI page that will be available at:
http://localhost:4000/status. The port used can be configured from a
configuration key (http_port). At the moment the UI will show the list
of failed jobs only. The idea of this commit is to propose a starting
point for an admin UI and discuss how can it be improved.

The changes are:

* a simple admin UI available at http://localhost:4000/status
* a new http API available at http://localhost:4000/api/
* at the moment the API exposes only one endpoint that returns the list
of failed jobs (/api/failed_jobs)
* update the README.md